Remove github support + support local git repos
The github support is quite poor (little error handling leading to
confusing exceptions) and not maintained, so we're getting rid of it.
Anyway, it should not be monolithe's job to setup the build environment
or to handle deployment. This also simplifies the code.

Since we still want to be able to manage sdks generated from multiple
branches of a given spec repo, we added local git repo support.